#470 : Renamed NautilusQueryCompiler.hpp/cpp to QueryCompiler.hpp/cpp
